feat(form): Refactor `InputCheckbox` to extend `BaseInput` and implement rendering logic.
